requirements.txt
Pillow==9.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
flask
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-hub==0.7.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.2.0
beautifulsoup4
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
flask-cors
nsfw_detector-1.1.1-py3-none-any.whl
yarg==0.1.9
gunicorn